[Refactor] Rename grid to module_grid and genelpa to module_genelpa (#6386)
* Rename grid to module_grid

* Rename genelpa to module_genelpa

* Fix cmake
