+++ source/source_base/CMakeLists.txt
@@ -85,7 +85,7 @@ if(BUILD_TESTING)
     add_subdirectory(kernels/test)
     add_subdirectory(module_mixing/test)
     add_subdirectory(module_device/test)
-    add_subdirectory(grid/test)
+    add_subdirectory(module_grid/test)
     if (USE_ABACUS_LIBM)
       add_subdirectory(libm/test)
     endif()
