+++ source/source_base/module_grid/partition.cpp
@@ -1,4 +1,4 @@
-#include "source_base/grid/partition.h"
+#include "source_base/module_grid/partition.h"
 #include "source_base/constants.h"
 
 #include <cmath>
