+++ source/source_hsolver/diago_elpa_native.cpp
@@ -7,7 +7,7 @@
 #include "source_base/scalapack_connector.h"
 #include "source_base/timer.h"
 #include "source_base/tool_quit.h"
-#include "source_hsolver/genelpa/elpa_new.h"
+#include "source_hsolver/module_genelpa/elpa_new.h"
 #include "omp.h"
 
 namespace hsolver
